Push notifications refactor (#243)
* Start refactoring the notifications

* break out notifications and milestones into their own functions

* Lots more refactoring

* more fixes

* Notifications should be hooked up end to end. Next...milestones!!!

* Push notifications hooked up in milestones!

* lint

* Refactor to make writing tests easier and add one unit test

* Unnecessary part of the test

* A mess but things are starting to work

* start working on milestones

* milestones and user subscriptions working

* Fix test and lint

* Move everything in notifications into a single tx block and fix a bug

* Undo this

* batch sending push notifications to SNS

* don't process old blocks?

* More tweaks to endpoints

* lint
