+++ identity-service/src/awsSNS.js
@@ -1,4 +1,6 @@
 const config = require('./config')
+const models = require('./models')
+const { logger } = require('./logging')
 
 // AWS SNS init
 const AWS = require('aws-sdk')
@@ -7,6 +9,10 @@ const sns = new AWS.SNS({
   secretAccessKey: config.get('awsSecretAccessKey'),
   region: 'us-west-1'
 })
+
+// TODO (DM) - move this into redis
+const PUSH_NOTIFICATION_SLICE_SIZE = 20
+let PUSH_NOTIFICATIONS_BUFFER = []
 
 // the aws sdk doesn't like when you set the function equal to a variable and try to call it
 // eg. const func = sns.<functionname>; func() returns an error, so util.promisify doesn't work
@@ -21,10 +27,91 @@ function _promisifySNS (functionName) {
   }
 }
 
+/**
+ * Formats a push notification in a way that's compatible with SNS
+ * @param {String} title title of push notification
+ * @param {String} body body of push notification
+ * @param {String} targetARN aws arn address for device
+ *                           `arn:aws:sns:us-west-1:<id>:endpoint/APNS/<namespace>/<uuid>`
+ * @param {Boolean=True} playSound should play a sound when it's sent
+ */
+function _formatIOSMessage (message, targetARN, playSound = true) {
+  let type = null
+  if (targetARN.includes('APNS_SANDBOX')) type = 'APNS_SANDBOX'
+  else if (targetARN.includes('APNS')) type = 'APNS'
+
+  const jsonMessage = {
+    'default': 'You have new notifications in Audius!'
+  }
+
+  // set iphone specific properties here
+  if (type) {
+    let apnsConfig = {
+      'aps': {
+        'alert': `${message}`
+        // keeping these properties here so we can use them if we want to
+        // "alert": {
+        //   "title" : `${title}`,
+        //   "body" : `${body}`
+        // },
+        // "badge": 19
+      }
+    }
+
+    jsonMessage[type] = JSON.stringify(apnsConfig)
+    if (playSound) jsonMessage[type].sound = 'default'
+  }
+
+  var params = {
+    Message: JSON.stringify(jsonMessage), /* required */
+    MessageStructure: 'json',
+    TargetArn: targetARN
+  }
+
+  return params
+}
+
 const listEndpointsByPlatformApplication = _promisifySNS('listEndpointsByPlatformApplication')
 const createPlatformEndpoint = _promisifySNS('createPlatformEndpoint')
+const publishPromisified = _promisifySNS('publish')
+const deleteEndpoint = _promisifySNS('deleteEndpoint')
+
+async function publish (message, userId, tx, playSound = true) {
+  const deviceInfo = await models.NotificationDeviceToken.findOne({ where: { userId }, transaction: tx })
+  if (!deviceInfo) return
+
+  let formattedMessage = null
+  if (deviceInfo.deviceType === 'ios') {
+    formattedMessage = _formatIOSMessage(message, deviceInfo.awsARN, playSound)
+  }
+
+  if (formattedMessage) {
+    logger.debug('AWS SNS formattedMessage', formattedMessage)
+    PUSH_NOTIFICATIONS_BUFFER.push(formattedMessage)
+  } else return null
+}
+
+async function drainPublishedMessages () {
+  try {
+    // TODO (DM) - batch this. DON'T DO Promise.all()
+    for (let i = 0; i < PUSH_NOTIFICATIONS_BUFFER.length; i += PUSH_NOTIFICATION_SLICE_SIZE) {
+      const pushNotifsSlice = PUSH_NOTIFICATIONS_BUFFER.slice(i, i + PUSH_NOTIFICATION_SLICE_SIZE)
+      await Promise.all(pushNotifsSlice.map((notification) => {
+        publishPromisified(notification)
+      }))
+    }
+
+    PUSH_NOTIFICATIONS_BUFFER = []
+  } catch (e) {
+    logger.error('Error sending push notification', e)
+    throw e
+  }
+}
 
 module.exports = {
   listEndpointsByPlatformApplication,
-  createPlatformEndpoint
+  createPlatformEndpoint,
+  publish,
+  deleteEndpoint,
+  drainPublishedMessages
 }
