+++ identity-service/src/notifications/constants.js
@@ -1,0 +1,81 @@
+const notificationTypes = Object.freeze({
+  Follow: 'Follow',
+  Repost: {
+    base: 'Repost',
+    track: 'RepostTrack',
+    album: 'RepostAlbum',
+    playlist: 'RepostPlaylist'
+  },
+  Favorite: {
+    base: 'Favorite',
+    track: 'FavoriteTrack',
+    album: 'FavoriteAlbum',
+    playlist: 'FavoritePlaylist'
+  },
+  Create: {
+    base: 'Create',
+    track: 'CreateTrack',
+    album: 'CreateAlbum',
+    playlist: 'CreatePlaylist'
+  },
+  Milestone: 'Milestone',
+  MilestoneFollow: 'MilestoneFollow',
+  MilestoneRepost: 'MilestoneRepost',
+  MilestoneFavorite: 'MilestoneFavorite',
+  MilestoneListen: 'MilestoneListen',
+  Announcement: 'Announcement'
+})
+
+const actionEntityTypes = Object.freeze({
+  User: 'User',
+  Track: 'Track',
+  Album: 'Album',
+  Playlist: 'Playlist'
+})
+
+const dayInHours = 24
+const weekInHours = 168
+const notificationJobType = 'notificationProcessJob'
+
+// TODO (DM) - unify this with the email messages
+const pushNotificationMessagesMap = {
+  [notificationTypes.Favorite.base] (notification) {
+    const [user] = notification.users
+    return `${user.name} favorited your ${notification.entity.type.toLowerCase()} ${notification.entity.name}`
+  },
+  [notificationTypes.Repost.base] (notification) {
+    const [user] = notification.users
+    return `${user.name} reposted your ${notification.entity.type.toLowerCase()} ${notification.entity.name}`
+  },
+  [notificationTypes.Follow] (notification) {
+    const [user] = notification.users
+    return `${user.name} followed you`
+  },
+  [notificationTypes.Announcement.base] (notification) {
+    return notification.text
+  },
+  [notificationTypes.Milestone] (notification) {
+    if (notification.entity) {
+      const entity = notification.entity.type.toLowerCase()
+      return `Your ${entity} ${notification.entity.name} has reached over ${notification.value} ${notification.achievement}s`
+    } else {
+      return `You have reached over ${notification.value} Followers `
+    }
+  },
+  [notificationTypes.Create.base] (notification) {
+    const [user] = notification.users
+    if (notification.entity.type === actionEntityTypes.Track && !isNaN(notification.entity.count) && notification.entity.count > 1) {
+      return `${user.name} released ${notification.entity.count} new ${notification.entity.type}`
+    }
+    return `${user.name} released a new ${notification.entity.type} ${notification.entity.name}`
+  }
+}
+
+module.exports = {
+  notificationTypes,
+  actionEntityTypes,
+  dayInHours,
+  weekInHours,
+  notificationJobType,
+  pushNotificationMessagesMap
+}
