+++ identity-service/src/notifications/formatNotificationMetadata.js
@@ -144,3 +144,4 @@ function formatNotificationProps (notifications, metadata) {
 }
 
 module.exports = formatNotificationProps
+module.exports.notificationResponseMap = notificationResponseMap
