+++ identity-service/test/index.js
@@ -1,3 +1,4 @@
 require('./expressAppTest')
 require('./authenticationTest')
 require('./relayTest')
+require('./notificationsTest')
